+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/DeadlockWatchdogThread.java
@@ -1,17 +1,29 @@
 package io.github.pr0methean.betterrandom;
 
-import io.github.pr0methean.betterrandom.util.LogPreFormatter;
 import io.github.pr0methean.betterrandom.util.LooperThread;
 import java.lang.management.ManagementFactory;
 import java.lang.management.ThreadInfo;
 import java.lang.management.ThreadMXBean;
-import java.util.logging.Level;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 @SuppressWarnings("unused") // intermittently needed for debugging
 public class DeadlockWatchdogThread extends LooperThread {
 
+  private static final class StackTraceHolder extends Throwable {
+    public StackTraceHolder(String name, StackTraceElement[] stackTrace) {
+      super(name, null, false, true);
+      setStackTrace(stackTrace);
+    }
+
+    @Override public synchronized Throwable fillInStackTrace() {
+      // No-op: we only use the stack trace that's in our constructor parameter
+      return this;
+    }
+  }
+
   private static final ThreadMXBean THREAD_MX_BEAN = ManagementFactory.getThreadMXBean();
-  private static final LogPreFormatter LOG = new LogPreFormatter(DeadlockWatchdogThread.class);
+  private static final Logger LOG = LoggerFactory.getLogger(DeadlockWatchdogThread.class);
   private static final int MAX_STACK_DEPTH = 20;
   private static final int DEADLOCK_STATUS = 0xDEAD10CC;
   private static DeadlockWatchdogThread INSTANCE = new DeadlockWatchdogThread();
@@ -44,14 +56,11 @@ public class DeadlockWatchdogThread extends LooperThread {
       throws InterruptedException {
     sleep(60_000);
     boolean deadlockFound = false;
-    final Level logLevel;
     long[] threadsOfInterest = THREAD_MX_BEAN.findDeadlockedThreads();
     if ((threadsOfInterest != null) && (threadsOfInterest.length > 0)) {
       LOG.error("DEADLOCKED THREADS FOUND");
-      logLevel = Level.SEVERE;
       deadlockFound = true;
     } else {
-      logLevel = Level.INFO;
       threadsOfInterest = THREAD_MX_BEAN.getAllThreadIds();
       if (threadsOfInterest.length <= 0) {
         LOG.error("ThreadMxBean didn't return any thread IDs");
@@ -60,9 +69,13 @@ public class DeadlockWatchdogThread extends LooperThread {
     }
     for (final long id : threadsOfInterest) {
       final ThreadInfo threadInfo = THREAD_MX_BEAN.getThreadInfo(id, MAX_STACK_DEPTH);
-      LOG.format(logLevel, 0, threadInfo.getThreadName());
       final StackTraceElement[] stackTrace = threadInfo.getStackTrace();
-      LOG.logStackTrace(logLevel, stackTrace);
+      Throwable t = new StackTraceHolder(threadInfo.getThreadName(), stackTrace);
+      if (deadlockFound) {
+        LOG.error("A deadlocked thread:", t);
+      } else {
+        LOG.info("A running thread:", t);
+      }
     }
     if (deadlockFound) {
       // Fail fast if current context allows