+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/BaseRandomTest.java
@@ -115,7 +115,6 @@ public abstract class BaseRandomTest {
   }
 
   @BeforeClass public void setUp() {
-    RandomSeederThread.setLoggingEnabled(false);
   }
 
   @Test public void testAllPublicConstructors()